fix(windows): make mouse move common between llhook and interception
