+++ src/oskbd/windows/llhook.rs
@@ -5,7 +5,6 @@
 
 use std::cell::Cell;
 use std::io;
-use std::mem::*;
 use std::{mem, ptr};
 
 use winapi::ctypes::*;
@@ -201,13 +200,7 @@ impl KbdOut {
     }
 
     pub fn move_mouse(&mut self, direction: MoveDirection, distance: u16) -> Result<(), io::Error> {
-        log::debug!("move mouse: {direction:?} {distance:?}");
-        match direction {
-            MoveDirection::Up => move_mouse(0, -i32::from(distance)),
-            MoveDirection::Down => move_mouse(0, i32::from(distance)),
-            MoveDirection::Left => move_mouse(-i32::from(distance), 0),
-            MoveDirection::Right => move_mouse(i32::from(distance), 0),
-        }
+        super::move_mouse(direction, distance);
         Ok(())
     }
 }
@@ -275,28 +268,6 @@ fn hscroll(direction: MWheelDirection, distance: u16) {
         *inputs[0].u.mi_mut() = m_input;
         SendInput(1, inputs.as_mut_ptr(), mem::size_of::<INPUT>() as _);
     }
-}
-
-// Taken from Enigo: https://github.com/enigo-rs/enigo
-fn mouse_event(flags: u32, data: u32, dx: i32, dy: i32) {
-    let mut input = INPUT {
-        type_: INPUT_MOUSE,
-        u: unsafe {
-            transmute(MOUSEINPUT {
-                dx,
-                dy,
-                mouseData: data,
-                dwFlags: flags,
-                time: 0,
-                dwExtraInfo: 0,
-            })
-        },
-    };
-    unsafe { SendInput(1, &mut input as LPINPUT, size_of::<INPUT>() as c_int) };
-}
-
-fn move_mouse(x: i32, y: i32) {
-    mouse_event(MOUSEEVENTF_MOVE, 0, x, y);
 }
 
 fn key_input_from_event(key: InputEvent) -> KEYBDINPUT {
