+++ src/oskbd/windows/mod.rs
@@ -1,8 +1,11 @@
 use std::mem;
 
+use winapi::ctypes::*;
 use winapi::um::winuser::*;
 
 use encode_unicode::CharExt;
+
+use crate::custom_action::*;
 
 #[cfg(not(feature = "interception_driver"))]
 mod llhook;
@@ -43,4 +46,36 @@ fn send_uc(c: char, up: bool) {
             mem::size_of::<INPUT>() as _,
         );
     }
+}
+
+fn move_mouse(direction: MoveDirection, distance: u16) {
+    log::debug!("move mouse: {direction:?} {distance:?}");
+    match direction {
+        MoveDirection::Up => move_mouse_xy(0, -i32::from(distance)),
+        MoveDirection::Down => move_mouse_xy(0, i32::from(distance)),
+        MoveDirection::Left => move_mouse_xy(-i32::from(distance), 0),
+        MoveDirection::Right => move_mouse_xy(i32::from(distance), 0),
+    }
+}
+
+fn move_mouse_xy(x: i32, y: i32) {
+    mouse_event(MOUSEEVENTF_MOVE, 0, x, y);
+}
+
+// Taken from Enigo: https://github.com/enigo-rs/enigo
+fn mouse_event(flags: u32, data: u32, dx: i32, dy: i32) {
+    let mut input = INPUT {
+        type_: INPUT_MOUSE,
+        u: unsafe {
+            mem::transmute(MOUSEINPUT {
+                dx,
+                dy,
+                mouseData: data,
+                dwFlags: flags,
+                time: 0,
+                dwExtraInfo: 0,
+            })
+        },
+    };
+    unsafe { SendInput(1, &mut input as LPINPUT, mem::size_of::<INPUT>() as c_int) };
 }
